add common.go
