+++ wgpu/wgpu.go
@@ -128,68 +128,3 @@ func (p *RenderBundle) Release()    { C.wgpuRenderBundleRelease(p.ref) }
 func (p *Sampler) Release()         { C.wgpuSamplerRelease(p.ref) }
 func (p *ShaderModule) Release()    { C.wgpuShaderModuleRelease(p.ref) }
 func (p *TextureView) Release()     { C.wgpuTextureViewRelease(p.ref) }
-
-// common types
-
-type Limits struct {
-	MaxTextureDimension1D                     uint32
-	MaxTextureDimension2D                     uint32
-	MaxTextureDimension3D                     uint32
-	MaxTextureArrayLayers                     uint32
-	MaxBindGroups                             uint32
-	MaxBindingsPerBindGroup                   uint32
-	MaxDynamicUniformBuffersPerPipelineLayout uint32
-	MaxDynamicStorageBuffersPerPipelineLayout uint32
-	MaxSampledTexturesPerShaderStage          uint32
-	MaxSamplersPerShaderStage                 uint32
-	MaxStorageBuffersPerShaderStage           uint32
-	MaxStorageTexturesPerShaderStage          uint32
-	MaxUniformBuffersPerShaderStage           uint32
-	MaxUniformBufferBindingSize               uint64
-	MaxStorageBufferBindingSize               uint64
-	MinUniformBufferOffsetAlignment           uint32
-	MinStorageBufferOffsetAlignment           uint32
-	MaxVertexBuffers                          uint32
-	MaxBufferSize                             uint64
-	MaxVertexAttributes                       uint32
-	MaxVertexBufferArrayStride                uint32
-	MaxInterStageShaderComponents             uint32
-	MaxInterStageShaderVariables              uint32
-	MaxColorAttachments                       uint32
-	MaxColorAttachmentBytesPerSample          uint32
-	MaxComputeWorkgroupStorageSize            uint32
-	MaxComputeInvocationsPerWorkgroup         uint32
-	MaxComputeWorkgroupSizeX                  uint32
-	MaxComputeWorkgroupSizeY                  uint32
-	MaxComputeWorkgroupSizeZ                  uint32
-	MaxComputeWorkgroupsPerDimension          uint32
-
-	MaxPushConstantSize uint32
-}
-
-type Color struct {
-	R, G, B, A float64
-}
-
-type Origin3D struct {
-	X, Y, Z uint32
-}
-
-type ImageCopyTexture struct {
-	Texture  *Texture
-	MipLevel uint32
-	Origin   Origin3D
-	Aspect   TextureAspect
-}
-
-type TextureDataLayout struct {
-	Offset       uint64
-	BytesPerRow  uint32
-	RowsPerImage uint32
-}
-
-type Extent3D struct {
-	Width              uint32
-	Height             uint32
-	DepthOrArrayLayers uint32
-}
